Refactor JpegR; push common into UltraHdr
